% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_level1.R
\name{corr_level1}
\alias{corr_level1}
\title{Calculate and Format Correlation Matrix For Level 1 Data}
\usage{
corr_level1(
  data,
  var_list,
  type = c("pearson", "spearman", "kendall"),
  digits = 3,
  decimal.mark = ".",
  triangle = c("both", "upper", "lower"),
  use = c("all.obs", "complete.obs", "pairwise.complete.obs"),
  show_significance = TRUE,
  replace_diagonal = TRUE,
  replacement = NA,
  lead.decimal = FALSE,
  var_labels = NULL,
  mean = TRUE,
  sd = TRUE,
  alpha = TRUE
)
}
\arguments{
\item{data}{A data frame containing the variables specified in \code{var_list}.}

\item{var_list}{A named list of variables where each entry is either a vector of
column indices or column names from the \code{data} data frame.}

\item{type}{The type of correlation to compute. Options are
"pearson" (default), "spearman", or "kendall".}

\item{digits}{Integer indicating the number of decimal places to be used for
formatting the correlation values. Default is 3.}

\item{decimal.mark}{Character to be used as the decimal mark. Default is ".".}

\item{triangle}{A character string indicating which part of the correlation
matrix to display. Options are "both", "upper", or "lower" (default).}

\item{use}{A character string indicating how to handle missing values.
Options are "all.obs", "complete.obs" (listwise deletion), or
"pairwise.complete.obs" (pairwise deletion - default).}

\item{show_significance}{Logical indicating whether to display
significance levels. Default is TRUE.}

\item{replace_diagonal}{Logical indicating whether to replace the
diagonal values. Default is TRUE.}

\item{replacement}{Value to replace the diagonal values if
\code{replace_diagonal} is TRUE. Default is NA.}

\item{lead.decimal}{Logical indicating whether to keep leading zeros
before the decimal point. Default is FALSE.}

\item{var_labels}{An optional vector of variable labels to use in the output.
Default is NULL.}

\item{mean}{Logical indicating whether to include the mean of the variables.
Default is TRUE.}

\item{sd}{Logical indicating whether to include the standard deviation of
the variables. Default is TRUE.}

\item{alpha}{Logical indicating whether to include Cronbach's alpha for
the variables. Default is TRUE.}
}
\value{
A formatted correlation matrix as a data frame, which can be exported
as a CSV file.
}
\description{
This function calculates the correlation matrix for a given list of variables,
formats the results, and includes additional metrics such as mean, standard deviation,
and Cronbach's alpha if requested.
}
\examples{
# Example usage
data <- teamstate
\dontrun{
var_list <- list(
  var1 = 3,
  ver2 = 4,
  var2 = 5:14,
  var3 = 15:24,
  var4 = 25:28
)
result <- corr_level1(data,
                      var_list,
                      var_labels = c("Gender", "Age", "Positive affect",
                      "Negative affect", "Psychological safety"))
print(result)
write.csv(result, file = "Correlation table individual member states.csv", row.names = TRUE)
}

}
\seealso{
\code{\link{cor}}, \code{\link{cor.test}}
}
