% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_level2.R
\name{corr_level2}
\alias{corr_level2}
\title{Calculate and Format Correlation Matrix for Level 2 Data}
\usage{
corr_level2(
  data,
  var_list,
  groupid,
  type = c("pearson", "spearman", "kendall"),
  digits = 3,
  decimal.mark = ".",
  triangle = c("lower", "both", "upper"),
  use = "pairwise.complete.obs",
  show_significance = TRUE,
  replace_diagonal = TRUE,
  replacement = NA,
  lead.decimal = TRUE,
  var_labels = NULL,
  mean = TRUE,
  sd = TRUE,
  alpha = TRUE,
  rwg = TRUE,
  rwg_scale = 7,
  rwg_method = c("mean", "median")
)
}
\arguments{
\item{data}{A data frame containing the variables specified in \code{var_list}.}

\item{var_list}{A named list of variables where each entry is either a vector of
column indices or column names from the \code{data} data frame.}

\item{groupid}{The column name in \code{data} that indicates group membership.}

\item{type}{The type of correlation to compute. Options are
"pearson" (default), "spearman", or "kendall".}

\item{digits}{Integer indicating the number of decimal places to be used for
formatting the correlation values. Default is 3.}

\item{decimal.mark}{Character to be used as the decimal mark. Default is ".".}

\item{triangle}{A character string indicating which part of the correlation
matrix to display. Options are "both", "upper", or "lower" (default is "lower").}

\item{use}{A character string indicating how to handle missing values.
Options are "all.obs", "complete.obs" (listwise deletion), or
"pairwise.complete.obs" (pairwise deletion - default).}

\item{show_significance}{Logical indicating whether to display
significance levels. Default is TRUE.}

\item{replace_diagonal}{Logical indicating whether to replace the
diagonal values. Default is TRUE.}

\item{replacement}{Value to replace the diagonal values if
\code{replace_diagonal} is TRUE. Default is NA.}

\item{lead.decimal}{Logical indicating whether to keep leading zeros
before the decimal point. Default is TRUE.}

\item{var_labels}{An optional vector of variable labels to use in the output.
Default is NULL.}

\item{mean}{Logical indicating whether to include the mean of
the aggregated variables at Level 2 (e.g., group level).
Default is TRUE.}

\item{sd}{Logical indicating whether to include the standard deviation of
the aggregated variables at Level 2 (e.g., group level). Default is TRUE.}

\item{alpha}{Logical indicating whether to include Cronbach's alpha for
the variables, which is calculated using items at Level 1. Default is TRUE.}

\item{rwg}{Logical indicating whether to include the mean or median of rwg.j,
which is calculated by function multilevel::rwg.j.
Default is TRUE.}

\item{rwg_scale}{The scale for calculating rwg.j, representing the number of
response options (e.g., 1 = Strongly Disagree to 5 = Strongly Agree). Default is 5.}

\item{rwg_method}{The method for calculating rwg.j.
Options are "mean" (default) or "median".}
}
\value{
A formatted correlation matrix as a data frame, which can be exported
as a CSV file.
}
\description{
This function calculates the correlation matrix for a given list of Level 2
variables using the aggregated variables (e.g., group level). The function
formats the results, and includes additional metrics such as mean, standard deviation,
Cronbach's alpha, and mean or median rwg.j if requested.
}
\examples{
\dontrun{
file_path <- system.file("extdata", "teamstate.csv", package = "mltable")
teamstate <- read.csv(file_path)
var_list2 <- list(
  var1 = 5:14,
  var2 = 15:24,
  var3 = 25:28
)
result <- corr_level2(data = teamstate,
                      var_list = var_list2,
                      groupid = "Team",
                      var_labels = c("Team PA", "Team NA", "Team Psychological Safety"),
                      rwg_scale = 7)
print(result)
}

}
\seealso{
\code{\link{cor}}, \code{\link{cor.test}}
}
